Beta-diversity

Network β-diversity

Measures of β-diversity work by first calculating the unique/shared items (using the βs, βos, and βwn functions), then passing on these arguments to one of the KGLXX functions to return a (dis)similarity score. The KGL functions are named for Koleff, Gaston, and Lennon – the number of each function matches the number in Table 1.

β-diversity components

The package implements functions for the βs, βos, and βwn components of network dissimilarity. In the original publication, we also described βst, which was the proprotion of dissimilarity due to species turnover, and defined as βst = βwn - βos for measures of dissimilarity bounded between 0 and 1. After discussing with colleagues and considering our own use-cases, it appears that the interpretation of βst is not always straightforward, and so we have decided to exclude it form the available functions.

EcologicalNetworks.βs — Function.

βs(X::T, Y::T) where {T<:BinaryNetwork}

Components of β-diversity as measured on species.

References

Koleff, P., Gaston, K.J., Lennon, J.J., 2003. Measuring beta diversity for presence–absence data. Journal of Animal Ecology 72, 367–382. https://doi.org/10.1046/j.1365-2656.2003.00710.x

source

EcologicalNetworks.βos — Function.

βos(X::T, Y::T) where {T<:BipartiteNetwork}

Overlapping species (bipartite)

References

Canard, E.F., Mouquet, N., Mouillot, D., Stanko, M., Miklisova, D., Gravel, D.,

Empirical evaluation of neutral interactions in host-parasite networks.

The American Naturalist 183, 468–479. https://doi.org/10.1086/675363

Poisot, T., Canard, E., Mouillot, D., Mouquet, N., Gravel, D., 2012. The dissimilarity of species interaction networks. Ecol. Lett. 15, 1353–1361. https://doi.org/10.1111/ele.12002

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

source

βos(X::T, Y::T) where {T<:UnipartiteNetwork}

Overlapping species (unipartite)

References

Canard, E.F., Mouquet, N., Mouillot, D., Stanko, M., Miklisova, D., Gravel, D.,

Empirical evaluation of neutral interactions in host-parasite networks.

The American Naturalist 183, 468–479. https://doi.org/10.1086/675363

Poisot, T., Canard, E., Mouillot, D., Mouquet, N., Gravel, D., 2012. The dissimilarity of species interaction networks. Ecol. Lett. 15, 1353–1361. https://doi.org/10.1111/ele.12002

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

source

EcologicalNetworks.βwn — Function.

βwn(X::T, Y::T) where {T<:BinaryNetwork}

Whole network

References

Canard, E.F., Mouquet, N., Mouillot, D., Stanko, M., Miklisova, D., Gravel, D.,

Empirical evaluation of neutral interactions in host-parasite networks.

The American Naturalist 183, 468–479. https://doi.org/10.1086/675363

Poisot, T., Canard, E., Mouillot, D., Mouquet, N., Gravel, D., 2012. The dissimilarity of species interaction networks. Ecol. Lett. 15, 1353–1361. https://doi.org/10.1111/ele.12002

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

source

β-diversity measures

EcologicalNetworks.KGL01 — Function.

KGL01(S::T)

Whittaker

source

Basic operations on networks

Internally, the functions for β-diversity rely on the usual operations on sets. The act of combining two networks, for example, is a union operation.

Base.setdiff — Function.

setdiff(X::T, Y::T) where {T<:BipartiteNetwork}

Difference between two bipartite networks. This operation is sensitive to the order of arguments, as the resulting network will have the species present in the first network (and their interactions) only.

source

setdiff(X::T, Y::T) where {T<:UnipartiteNetwork}

Difference between two unipartite networks. This operation is sensitive to the order of arguments, as the resulting network will have the species present in the first network (and their interactions) only.

source

Base.union — Function.

union(X::T, Y::T) where {T<:BipartiteNetwork}

Union of two bipartite networks – interactions and species which are present in either networks are also present in the final network.

source

union(X::T, Y::T) where {T<:UnipartiteNetwork}

Union of two unipartite networks – interactions and species which are present in either networks are also present in the final network.

source

Base.intersect — Function.

intersect(X::T, Y::T) where {T<:BipartiteNetwork}

Intersect between two bipartite networks. The resulting network has the species and interactions common to both networks. This can result in species being disconnected, if they are found in both networks but with no operations in common.

source

intersect(X::T, Y::T) where {T<:UnipartiteNetwork}

Intersect between two unipartite networks. The resulting network has the species and interactions common to both networks. This can result in species being disconnected, if they are found in both networks but with no operations in common.

source