Internals

Base.:! — Method

Base.:!{T<:DeterministicNetwork}(N::T)

Returns the inverse of a binary network – interactions that were false become true, and conversely.

Base.:+ — Method

+(n1::T, n2::T) where {T <: BipartiteQuantitativeNetwork}

Adds two quantitative bipartite networks. TODO

Base.convert — Method

convert(::Type{AbstractUnipartiteNetwork}, N::AbstractBipartiteNetwork)

Projects any bipartite network in its unipartite representation. This function will call the correct type-to-type convert function depending on the type of the input network.

The type to be converted to must be AbstractUnipartiteNetwork – for example, converting a bipartite probabilistic network to a probabilistic unipartite network is not a meaningful operation.

Base.convert — Method

convert(::Type{BinaryNetwork}, N::QuantitativeNetwork)

Projects any bipartite network in its unipartite representation. This function will call the correct type-to-type convert function depending on the type of the input network.

This function does not work for probabilistic networks. The operation of generating a deterministic network from a probabilistic one is different from a simple conversion: it can be done either through random draws, or by selecting only interactions with a probability greater than 0 (N>0.0 will do this).

Base.convert — Method

convert(::Type{BipartiteNetwork}, N::T) where {T <: BipartiteQuantitativeNetwork}

Convert a bipartite quantitative network to a bipartite binary network. This amounts to removing the quantitative information.

Base.convert — Method

convert(::Type{BipartiteNetwork}, N::T) where {T <: UnipartiteNetwork}

Projects a unipartite network (specifically, a BipartiteNetwork) to its bipartite representation. The following checks are performed.

First, the network cannot be degenerate, since species with no interactions cannot be assigned to a specific level. Second, the species cannot have both in and out degree. If these conditions are met, the bipartite network will be returned.

Base.convert — Method

convert(::Type{BipartiteProbabilisticNetwork}, N::T) where {T <: UnipartiteProbabilisticNetwork}

Projects a unipartite network (specifically, a BipartiteProbabilisticNetwork) to its bipartite representation. The following checks are performed.

First, the network cannot be degenerate, since species with no interactions cannot be assigned to a specific level. Second, the species cannot have both in and out degree. If these conditions are met, the bipartite network will be returned.

Base.convert — Method

convert(::Type{BipartiteQuantitativeNetwork}, N::T) where {T <: UnipartiteQuantitativeNetwork}

Projects a unipartite network (specifically, a BipartiteQuantitativeNetwork) to its bipartite representation. The following checks are performed.

First, the network cannot be degenerate, since species with no interactions cannot be assigned to a specific level. Second, the species cannot have both in and out degree. If these conditions are met, the bipartite network will be returned.

Base.convert — Method

convert(::Type{UnipartiteNetwork}, N::T) where {T <: BipartiteNetwork}

Projects a deterministic bipartite network in its unipartite representation.

Base.convert — Method

convert(::Type{UnipartiteNetwork}, N::T) where {T <: UnipartiteQuantitativeNetwork}

Convert a unipartite quantitative network to a unipartite binary network. This amounts to removing the quantitative information.

Base.convert — Method

convert(::Type{UnipartiteProbabilisticNetwork}, N::T) where {T <: BipartiteProbabilisticNetwork}

Projects a probabilistic bipartite network in its unipartite representation.

Base.convert — Method

convert(::Type{UnipartiteQuantitativeNetwork}, N::T) where {T <: BipartiteQuantitativeNetwork}

Projects a quantitative bipartite network in its unipartite representation.

Base.copy — Method

Creates a copy of a network – this returns an object with the same type, and the same content.

Base.getindex — Method

getindex{T}(N::AbstractBipartiteNetwork, sp1::Array{T}, sp2::Array{T})

TODO

Base.getindex — Method

getindex{T}(N::AbstractBipartiteNetwork, sp1::Array{T}, sp2::Array{T})

TODO

Base.getindex — Method

getindex{T}(N::AbstractBipartiteNetwork, sp::Array{T}, ::Colon)

TODO

Base.getindex — Method

getindex{T}(N::AbstractBipartiteNetwork, ::Colon, sp::Array{T})

TODO

Base.getindex — Method

getindex{T}(N::AbstractBipartiteNetwork, ::Colon, sp::T)

Gets the predecessors (i.e. species that interacts with / consume) of a focal species. This returns the list of species as a Set object, in which ordering is unimportant.

Base.getindex — Method

getindex(N::AbstractEcologicalNetwork, ::Colon, j::T)

Get the value of an interaction based on the position of the species.

Base.getindex — Method

getindex{T}(N::AbstractEcologicalNetwork, sp::T, ::Colon)

Gets the successors (i.e. species that are interacted with / consumed) of a focal species. This returns the list of species as a Set object, in which ordering is unimportant.

Base.getindex — Method

getindex(N::AbstractEcologicalNetwork, i::T, ::Colon)

Get the value of an interaction based on the position of the species.

Base.getindex — Method

getindex{T}(N::AbstractEcologicalNetwork, s1::T, s2::T)

Get the value of an interaction based on the name of the species. This is the recommended way to look for things in a network.

Base.getindex — Method

getindex(N::AbstractEcologicalNetwork, i::T, j::T)

Get the value of an interaction based on the position of the species.

Base.getindex — Method

getindex(N::AbstractEcologicalNetwork, i::T)

Get the value of an interaction based on the position of the species.

Base.getindex — Method

getindex{T}(N::AbstractUnipartiteNetwork, sp::Array{T})

Induce a unipartite network based on a list of species, all of which must be in the original network. This function takes a single argument (as opposed to two arrays, or an array and a colon) to ensure that the returned network is unipartite.

The network which is returned by this function may not have the species in the order specified by the user for performance reasons.

Base.intersect — Method

intersect(X::T, Y::T) where {T<:BipartiteNetwork}

Intersect between two bipartite networks. The resulting network has the species and interactions common to both networks. This can result in species being disconnected, if they are found in both networks but with no operations in common.

Base.intersect — Method

intersect(X::T, Y::T) where {T<:UnipartiteNetwork}

Intersect between two unipartite networks. The resulting network has the species and interactions common to both networks. This can result in species being disconnected, if they are found in both networks but with no operations in common.

Base.iterate — Method

Base.iterate(N::T) where {T <: AbstractEcologicalNetwork}

Returns the interactions in a network (this uses less memory than interactions but is a bit slower).

Base.iterate — Method

Base.iterate(N::T) where {T <: AbstractEcologicalNetwork}

Returns the interactions in a network (this uses less memory than interactions but is a bit slower).

Base.length — Method

Base.length(N::T) where {T <: AbstractEcologicalNetwork}

The length of a network is the number of non-zero elements in it.

Base.permutedims — Method

permutedims(N::AbstractBipartiteNetwork)

Tranposes the network and returns a copy

Base.rand — Method

rand(N::ProbabilisticNetwork, S::Tuple{Int64,Int64})

Generates a number of random deterministic networks based on a probabilistic network, and returns them as a matrix.

References

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

Base.rand — Method

rand(N::ProbabilisticNetwork)

Converts a probabilistic network into a deterministic one by performing random draws. All interactions are treated as independent Bernoulli events. Note that this network is not checked for degeneracy, i.e. species can end up with no interactions.

References

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

Base.rand — Method

rand(N::ProbabilisticNetwork, n::T) where {T<:Integer}

Generates a number of random deterministic networks based on a probabilistic network.

References

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

Base.setdiff — Method

setdiff(X::T, Y::T) where {T<:BipartiteNetwork}

Difference between two bipartite networks. This operation is sensitive to the order of arguments, as the resulting network will have the species present in the first network (and their interactions) only.

Base.setdiff — Method

setdiff(X::T, Y::T) where {T<:UnipartiteNetwork}

Difference between two unipartite networks. This operation is sensitive to the order of arguments, as the resulting network will have the species present in the first network (and their interactions) only.

Base.setindex! — Method

setindex!(N::T, A::Any, i::E, j::E) where {T <: AbstractEcologicalNetwork, E}

Changes the value of the interaction at the specificied position, where i and j are species names. Note that this operation changes the network.

Base.setindex! — Method

setindex!(N::T, A::K, i::E, j::E) where {T <: AbstractEcologicalNetwork, K <: _interaction_type(N), E <: Int}

Changes the value of the interaction at the specificied position, where i and j are species positions. Note that this operation changes the network.

Base.show — Method

show(io::IO, N::AbstractEcologicalNetwork)

Base.similar — Method

similar(N::T) where {T <: AbstractEcologicalNetwork}

Returns a network with the same species, but an empty interaction matrix. This is useful if you want to generate a "blank slate" for some analyses.

Base.size — Method

size(N::AbstractEcologicalNetwork, i::Int64)

Return the size of the adjacency matrix of an AbstractEcologicalNetwork object.

Base.size — Method

size(N::AbstractEcologicalNetwork)

Return the size of the adjacency matrix of an AbstractEcologicalNetwork object.

Base.sum — Method

sum(N::AbstractEcologicalNetwork; dims=nothing)

This function will return the sum of all interactions in the network. For quantitative networks, this is the sum of interaction strengths. For binary networks, this is the number of interactions. For probabilistic networks, this is the expected number of realized interactions.

Optionally, one can give the argument dims, simular to the native sum, which computes the sum of the interactions for the lower (dims=2) or higher (dims=1) level.

References

Delmas, E., Besson, M., Brice, M.-H., Burkle, L.A., Dalla Riva, G.V., Fortin, M.-J., Gravel, D., Guimarães, P.R., Hembry, D.H., Newman, E.A., Olesen, J.M., Pires, M.M., Yeakel, J.D., Poisot, T., 2018. Analysing ecological networks of species interactions. Biological Reviews 112540. https://doi.org/10.1111/brv.12433

Dunne, J.A., 2006. The Network Structure of Food Webs, in: Dunne, J.A., Pascual, M. (Eds.), Ecological Networks: Linking Structure and Dynamics. Oxford University Press, pp. 27–86.

Martinez, N.D., 1992. Constant Connectance in Community Food Webs. The American Naturalist 139, 1208–1218.

Base.union — Method

union(X::T, Y::T) where {T<:BipartiteNetwork}

Union of two bipartite networks – interactions and species which are present in either networks are also present in the final network.

Base.union — Method

union(X::T, Y::T) where {T<:UnipartiteNetwork}

Union of two unipartite networks – interactions and species which are present in either networks are also present in the final network.

EcologicalNetworks._additive_bernoulli_variance — Method

Variance of a series of additive Bernoulli events

f(p): ∑(p(1-p))

EcologicalNetworks._multiplicative_bernoulli_variance — Method

Variance of a series of multiplicative Bernoulli events

EcologicalNetworks._permute_motif — Method

Internal function

Returns all permutations of the adjacency matrix of a motif.

EcologicalNetworks._permute_motif — Method

Internal function

Returns all permutations of the adjacency matrix of a motif.

EcologicalNetworks._single_bernoulli_expectation — Method

Expected value of a single Bernoulli event

Simply f(p): p

EcologicalNetworks._single_bernoulli_variance — Method

Variance of a single Bernoulli event

f(p): p(1-p)

EcologicalNetworks._value_is_a_probability — Method

Quite crude way of checking that a number is a probability

The two steps are

The number should be of the Float64 type

The number should belong to [0.0,1.0]

EcologicalNetworks.degree_in — Method

degree_in(N::AbstractEcologicalNetwork)

EcologicalNetworks.degree_in_var — Method

Variance in the ingoing degree

degree_in_var(N::ProbabilisticNetwork)

EcologicalNetworks.degree_out — Method

degree_out(N::AbstractEcologicalNetwork)

EcologicalNetworks.degree_out_var — Method

Variance in the outgoing degree

degree_out_var(N::ProbabilisticNetwork)

EcologicalNetworks.mirror — Method

mirror(N::T) where {T <: UnipartiteNetwork}

Returns a mirrored version of the adjacency matrix, required for spectral radius.

EcologicalNetworks.nodf_axis — Method

NODF of a single axis

EcologicalNetworks.nodf_axis — Method

WNODF of a single axis

EcologicalNetworks.pdi — Method

pdi{T<:Number}(x::Array{T, 1})

This function will range the values of each row, so that the strongest link has a value of one. This works for deterministic and quantitative networks.

References

Poisot, T., Bever, J.D., Nemri, A., Thrall, P.H., Hochberg, M.E., 2011. A conceptual framework for the evolution of ecological specialisation. Ecol. Lett. 14, 841–851. https://doi.org/10.1111/j.1461-0248.2011.01645.x

Poisot, T., Canard, E., Mouquet, N., Hochberg, M.E., 2012. A comparative study of ecological specialization estimators. Methods in Ecology and Evolution 3, 537–544. https://doi.org/10.1111/j.2041-210X.2011.00174.x

EcologicalNetworks.η_axis — Method

Nestedness of a single axis (called internally by η)

Random.shuffle! — Method

shuffle!(N::BinaryNetwork; constraint::Symbol=:degree)

Shuffles interactions inside a network (the network is modified), under the following constraint:

:degree, which keeps the degree distribution intact

:generality, which keeps the out-degree distribution intact

:vulnerability, which keeps the in-degree distribution intact

:fill, which moves interactions around freely

The function will take two interactions, and swap the species establishing them. By repeating the process a large enough number of times, the resulting network should be relatively random. Note that this function will conserve the degree (when appropriate under the selected constraint) of every species. Calling the function will perform a single shuffle. If you want to repeat the shuffling a large enough number of times, you can use something like:

[shuffle!(n) for i in 1:10_000]

If the keyword arguments are invalid, the function will throw an ArgumentError.

References

Fortuna, M.A., Stouffer, D.B., Olesen, J.M., Jordano, P., Mouillot, D., Krasnov, B.R., Poulin, R., Bascompte, J., 2010. Nestedness versus modularity in ecological networks: two sides of the same coin? Journal of Animal Ecology 78, 811–817. https://doi.org/10.1111/j.1365-2656.2010.01688.x

Random.shuffle — Method

shuffle(N::BinaryNetwork; constraint::Symbol=:degree)

Return a shuffled copy of the network (the original network is not modified). See shuffle! for a documentation of the keyword arguments.

References

Fortuna, M.A., Stouffer, D.B., Olesen, J.M., Jordano, P., Mouillot, D., Krasnov, B.R., Poulin, R., Bascompte, J., 2010. Nestedness versus modularity in ecological networks: two sides of the same coin? Journal of Animal Ecology 78, 811–817. https://doi.org/10.1111/j.1365-2656.2010.01688.x