Motifs

Motif enumeration

List of canonical motifs

EcologicalNetworks.unipartitemotifs — Function.

unipartitemotifs()

The names of the motifs come from Stouffer et al. (2007) – especially Fig. 1, available online at http://rspb.royalsocietypublishing.org/content/274/1621/1931.figures-only

The motifs are returned as a named tuple, with every motif identified by its name in the original publication. The species are named :a. :b, and :c.

source

Motif counting

EcologicalNetworks.find_motif — Function.

find_motif(N::T1, m::T2) where {T1<:AbstractEcologicalNetwork, T2<:BinaryNetwork}

Returns an array of tuples, in which each tuple contains the species that are part of the motif. The length of the array gives the number of times the motif was found. For probabilistic networks, the tuple also contains the probability of observing the species in the correct conformation for the motif, as well as the variance.

References

Milo, R., Shen-Orr, S., Itzkovitz, S., Kashtan, N., Chklovskii, D., Alon, U.,

Network motifs: simple building blocks of complex networks. Science 298,

824–7. https://doi.org/10.1126/science.298.5594.824

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

source

Probabilistic case

EcologicalNetworks.expected_motif_count — Function.

expected_motif_count(s)

Get the expected number of motifs (and variance) from the output of find_motif on a probabilistic network.

References

Poisot, T., Cirtwill, A.R., Cazelles, K., Gravel, D., Fortin, M.-J., Stouffer, D.B., 2016. The structure of probabilistic networks. Methods in Ecology and Evolution 7, 303–312. https://doi.org/10.1111/2041-210X.12468

source