Structural network models

Structure of ecological networks is non-random. Network architecture can have a strong effect on important ecosystem properties (Mougi and Kondoh 2012, Thébault and Fontaine 2010). Many of the structural features of food-webs can be simulated using small number of simple rules. Despite this simplicity these models can often accurately reproduce some of the second order characteristics of empirical food-webs (Stouffer et al. 2005). These characteristics of phenomenological stochastic models allow for their wide applications e.g. to simulate biomass dynamics using dynamical models or study extinction cascades.

Mougi, A. and Kondoh, M. (2012) ‘Diversity of Interaction Types and

Ecological Community Stability’, Science, 337(6092), pp. 349–351. doi: 10.1126/science.1220529.

Thébault, E. and Fontaine, C. (2010) ‘Stability of Ecological Communities

and the Architecture of Mutualistic and Trophic Networks’, Science, 329(5993), pp. 853–856. doi: 10.1126/science.1188321.

Stouffer, D. B. et al. (2005) ‘Quantitative Patterns in the Structure

of Model and Empirical Food Webs’, Ecology, 86(5), pp. 1301–1311. doi: 10.1890/04-0957.

Many models with various interactions assignment algorithms have been proposed. EcologicalNetworks provides functions to generate random ecological networks of the UnipartiteNetwork type. Listed below are those most often used in ecological studies.

Cascade model

This model uses one abstract trophic trait. For any given consumer links can be randomly assigned to a resource species with the trait value smaller than that of a consumer.

EcologicalNetworks.cascademodel — Function

cascademodel(S::Int64, Co::Float64)

Return matrix of the type UnipartiteNetwork randomly assembled according to the cascade model for a given nuber of S and connectivity Co.

See also: nichemodel, mpnmodel, nestedhierarchymodel

References

Cohen, J.E., Newman, C.M., 1985. A stochastic theory of community food webs I. Models and aggregated data. Proceedings of the Royal Society of London. Series B. Biological Sciences 224, 421–448. https://doi.org/10.1098/rspb.1985.0042

cascademodel(N::T) where {T <: UnipartiteNetwork}

Applied to a UnipartiteNetwork return its randomized version.

References

Cohen, J.E., Newman, C.M., 1985. A stochastic theory of community food webs I. Models and aggregated data. Proceedings of the Royal Society of London. Series B. Biological Sciences 224, 421–448. https://doi.org/10.1098/rspb.1985.0042

cascademodel(S::Int64, L::Int64)

Number of links can be specified instead of connectance.

References

Cohen, J.E., Newman, C.M., 1985. A stochastic theory of community food webs I. Models and aggregated data. Proceedings of the Royal Society of London. Series B. Biological Sciences 224, 421–448. https://doi.org/10.1098/rspb.1985.0042

cascademodel(parameters::Tuple)

Parameters tuple can also be provided in the form (S::Int64, Co::Float64) or (S::Int64, L::Int64).

References

Cohen, J.E., Newman, C.M., 1985. A stochastic theory of community food webs I. Models and aggregated data. Proceedings of the Royal Society of London. Series B. Biological Sciences 224, 421–448. https://doi.org/10.1098/rspb.1985.0042

Niche model

Niche model extended cascade model by introducing ranges for each consumer. In this model consumers can predate on resources which trait values are within the predators' 'niche' range.

EcologicalNetworks.nichemodel — Function

nichemodel(S::Int64, L::Int64)

Return UnipartiteNetwork where resources are assign to consumers according to niche model for a network of S species and L links.

References

Williams, R., Martinez, N., 2000. Simple rules yield complex food webs. Nature 404, 180–183.

nichemodel(N::T) where {T <: UnipartiteNetwork}

Applied to empirical UnipartiteNetwork return its randomized version.

References

Williams, R., Martinez, N., 2000. Simple rules yield complex food webs. Nature 404, 180–183.

nichemodel(S::Int64, C::Float64)

References

Williams, R., Martinez, N., 2000. Simple rules yield complex food webs. Nature 404, 180–183.

nichemodel(parameters::Tuple)

Parameters tuple can also be provided in the form (Species::Int64, Co::Float64) or (Species::Int64, Int::Int64).

References

Williams, R., Martinez, N., 2000. Simple rules yield complex food webs. Nature 404, 180–183.

Nested hierarchy model

In order to reproduce more faithfully properties of complex and multidimensional natural nested hierarchy model tries to use simple rules to incorporate also the phylogenetic similarity in resource composition of predators.

EcologicalNetworks.nestedhierarchymodel — Function

nestedhierarchymodel(S::Int64, L::Int64)

Return UnipartiteNetwork where resources are assigned to consumers according to the nested hierarchy model for S species and L.

References

Cattin, M.-F., Bersier, L.-F., Banašek-Richter, C., Baltensperger, R., Gabriel, J.-P., 2004. Phylogenetic constraints and adaptation explain food-web structure. Nature 427, 835–839. https://doi.org/10.1038/nature02327

nestedhierarchymodel(S::Int64, Co::Float64)

Connectance can be provided instead of number of links.

References

Cattin, M.-F., Bersier, L.-F., Banašek-Richter, C., Baltensperger, R., Gabriel, J.-P., 2004. Phylogenetic constraints and adaptation explain food-web structure. Nature 427, 835–839. https://doi.org/10.1038/nature02327

nestedhierarchymodel(N::T) {T <: UnipartiteNetwork}

Applied to empirical UnipartiteNetwork return its randomized version.

References

Cattin, M.-F., Bersier, L.-F., Banašek-Richter, C., Baltensperger, R., Gabriel, J.-P., 2004. Phylogenetic constraints and adaptation explain food-web structure. Nature 427, 835–839. https://doi.org/10.1038/nature02327

nestedhierarchymodel(parameters::Tuple)

Parameters tuple can also be provided in the form (Species::Int64, Co::Float64) or (Species::Int64, Int::Int64).

References

Cattin, M.-F., Bersier, L.-F., Banašek-Richter, C., Baltensperger, R., Gabriel, J.-P., 2004. Phylogenetic constraints and adaptation explain food-web structure. Nature 427, 835–839. https://doi.org/10.1038/nature02327

Minimum potential niche model

This model attempts to explicitly simulate forbidden links in empirical food webs.

EcologicalNetworks.mpnmodel — Function

mpnmodel(S::Int64, Co::Float64, forbidden::Float64)

Return UnipartiteNetwork with links assigned according to minimum potential niche model for given number of S, connectivity Co and probability of forbidden link occurence.

References

Allesina, S., Alonso, D., Pascual, M., 2008. A General Model for Food Web Structure. Science 320, 658–661. https://doi.org/10.1126/science.1156269

mpnmodel(N::T) where {T<: UnipartiteNetwork}

Applied to UnipartiteNetwork return its randomized version.

References

Allesina, S., Alonso, D., Pascual, M., 2008. A General Model for Food Web Structure. Science 320, 658–661. https://doi.org/10.1126/science.1156269

mpnmodel(parameters::Tuple)

Parameters tuple can also be provided in the form (S::Int64, Co::Float64, forbidden::Float64).

References

Allesina, S., Alonso, D., Pascual, M., 2008. A General Model for Food Web Structure. Science 320, 658–661. https://doi.org/10.1126/science.1156269

mpnmodel(mpnmodel(S::Int64, L::Int64, forbidden::Float64))

Average number of links can be specified instead of connectance.

References

Allesina, S., Alonso, D., Pascual, M., 2008. A General Model for Food Web Structure. Science 320, 658–661. https://doi.org/10.1126/science.1156269