Plotting

Plotting functions are part of the EcologicalNetworksPlot, which requires Plot to work. EcologicalNetworksPlot can be installed from the central Julia package repository.

Initial layouts

EcologicalNetworksPlots.initial — Function.

initial(::Type{RandomInitialLayout}, N::T) where {T <: EcologicalNetworks.AbstractEcologicalNetwork}

Random disposition of nodes in the unit square. This is a good starting point for any force-directed layout.

initial(::Type{BipartiteInitialLayout}, N::T) where {T <: EcologicalNetworks.AbstractBipartiteNetwork}

Random disposition of nodes on two levels for bipartite networks.

initial(::Type{FoodwebInitialLayout}, N::T) where {T <: EcologicalNetworks.AbstractUnipartiteNetwork}

Random disposition of nodes on trophic levels for food webs. Note that the fractional trophic level is used, but the layout can be modified afterwards to use the continuous levels.

initial(::Type{CircularInitialLayout}, N::T) where {T <: EcologicalNetworks.AbstractEcologicalNetwork}

Random disposition of nodes on a circle. This is the starting point for circle-based layouts.

EcologicalNetworksPlots.RandomInitialLayout — Type.

RandomInitialLayout

This type is used to generate an initial layout, where the nodes are placed at random.

EcologicalNetworksPlots.BipartiteInitialLayout — Type.

BipartiteInitialLayout

This type is used to generate an initial bipartite layout, where the nodes are placed on two levels, but their horizontal position is random.

EcologicalNetworksPlots.FoodwebInitialLayout — Type.

FoodwebInitialLayout

This type is used to generate an initial layout, where the nodes are placed on their trophic levels, but their horizontal position is random.

EcologicalNetworksPlots.CircularInitialLayout — Type.

CircularInitialLayout

This type is used to generate an initial layout, where the nodes are placed at random along a circle.

Layouts

Force-directed layout

EcologicalNetworksPlots.ForceDirectedLayout — Type.

ForceDirectedLayout

Circular layout

EcologicalNetworksPlots.CircularLayout — Type.

CircularLayout

A circular layout has a single field, radius.

Static layouts

EcologicalNetworksPlots.NestedBipartiteLayout — Type.

NestedBipartiteLayout

Parameters are

align (whether the two levels should be centered together)

relative (whether the two levels should occupy a length equal to their relative richness)

spread (the distance between the two)

Note that to see the effect of spread, you may have to use aspectratio=1; if not, the spacing between levels will be determined by the dimensions of the plot.

Apply layout to network

EcologicalNetworksPlots.position! — Function.

position!(LA::ForceDirectedLayout, L::Dict{K,NodePosition}, N::T) where {T <: EcologicalNetworks.AbstractEcologicalNetwork} where {K}

One iteration of the force-directed layout routine. Because these algorithms can take some time to converge, it may be useful to stop every 500 iterations to have a look at the results.

position!(LA::CircularLayout, L::Dict{K,NodePosition}, N::T) where {T <: AbstractEcologicalNetwork} where {K}

Nodes will be positioned at equal distances along a circle, and nodes that are densely connected will be closer to one another. This is an efficient way to represent modular networks.

References

McGuffin, M.J., 2012. Simple algorithms for network visualization: A tutorial. Tsinghua Science and Technology 17, 383–398. https://doi.org/10.1109/TST.2012.6297585

Examples

Static layouts

Nested

Unes = web_of_life("M_SD_033")
I = initial(BipartiteInitialLayout, Unes)
position!(NestedBipartiteLayout(0.4), I, Unes)
plot(I, Unes, aspectratio=1)
scatter!(I, Unes, bipartite=true)

Circular

Unes = web_of_life("M_SD_033")
I = initial(CircularInitialLayout, Unes)
position!(CircularLayout(), I, Unes)
plot(I, Unes, aspectratio=1)
scatter!(I, Unes, bipartite=true)

Dynamic layouts

Force directed

Umod = web_of_life("M_PA_003")
I = initial(RandomInitialLayout, Umod)
for step in 1:4000
  position!(ForceDirectedLayout(2.5), I, Umod)
end
plot(I, Umod, aspectratio=1)
scatter!(I, Umod, bipartite=true)

Food web layout

Fweb = simplify(nz_stream_foodweb()[5])
I = initial(FoodwebInitialLayout, Fweb)
for step in 1:4000
  position!(ForceDirectedLayout(true, false, 2.5), I, Fweb)
end
plot(I, Fweb)
scatter!(I, Fweb)

Note that we can replace some properties of the nodes in the layout after the positioning algorithm occurred – so we can, for example, use the actual (instead of fractional) trophic level:

Fweb = simplify(nz_stream_foodweb()[5])
I = initial(FoodwebInitialLayout, Fweb)
for step in 1:4000
  position!(ForceDirectedLayout(true, false, 2.5), I, Fweb)
end
tl = trophic_level(Fweb)
for s in species(Fweb)
  I[s].y = tl[s]
end
plot(I, Fweb)
scatter!(I, Fweb)

Node properties

Color

Unes = web_of_life("M_SD_033")
I = initial(BipartiteInitialLayout, Unes)
position!(NestedBipartiteLayout(0.4), I, Unes)
plot(I, Unes, aspectratio=1)
scatter!(I, Unes, bipartite=true, nodefill=degree(Unes))

Size

Advanced examples

One important feature of the package is that the layout can contain more nodes than the network. For example, we can use this to our advantage, to represent species with a degree larger than 3 in red:

Umod = web_of_life("M_PA_003")
I = initial(RandomInitialLayout, Umod)
for step in 1:4000
  position!(ForceDirectedLayout(2.5), I, Umod)
end
plot(I, Umod, aspectratio=1)
scatter!(I, Umod)
N = convert(AbstractUnipartiteNetwork, convert(BinaryNetwork, Umod))
core3 = collect(keys(filter(p -> p.second ≥ 3, degree(N))))
plot!(I, N[core3], lc=:red)
scatter!(I, N[core3], mc=:red)

Heatmaps

Umod = web_of_life("M_PA_003")
heatmap(Umod)